# API/PruebaProyecto/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PruebaProyecto, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/CuartosFrios": {
      "get": {
        "tags": [
          "CuartosFrios"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "CuartosFrios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CuartosFrios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "CuartosFrios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CuartosFrios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "CuartosFrios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CuartosFrios/{id}": {
      "get": {
        "tags": [
          "CuartosFrios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CuartosFrios": {
      "required": [
        "cantidadActual",
        "capacidadMaxima"
      ],
      "type": "object",
      "properties": {
        "idCuarto": {
          "format": "int32",
          "type": "integer"
        },
        "capacidadMaxima": {
          "format": "int32",
          "type": "integer"
        },
        "cantidadActual": {
          "format": "int32",
          "type": "integer"
        },
        "capacidadDisponible": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      },
      "additionalProperties": false
    },
...
  }
}